+++ docs/examples/routing/blueprints.py
@@ -1,6 +1,6 @@
-from django.urls import include, path
+from django.urls import include
 
-from dmr.routing import Router, compose_blueprints
+from dmr.routing import Router, compose_blueprints, path
 from examples.using_controller.blueprints import (
     UserCreateBlueprint,
     UserListBlueprint,
@@ -24,3 +24,6 @@ router = Router(
 urlpatterns = [
     path(router.prefix, include((router.urls, 'server'), namespace='api')),
 ]
+
+# run: {"method": "get", "url": "/api/user/", "use_urlpatterns": true}  # noqa: ERA001, E501
+# run: {"method": "post", "url": "/api/user/", "body": {"email": "some@example.com", "age": 20}, "use_urlpatterns": true}  # noqa: ERA001, E501