+++ tests/test_unit/test_openapi/test_schema_snapshots.py
@@ -1,27 +1,27 @@
 import json
 from http import HTTPStatus
-from typing import Annotated, ClassVar, Literal
+from typing import Annotated, ClassVar
 
 import pydantic
-from django.urls import path
+from django.urls import path, re_path
 from syrupy.assertion import SnapshotAssertion
 
 from dmr import (
+    Blueprint,
     Body,
     Controller,
     Cookies,
-    FileMetadata,
+    Path,
     Query,
     ResponseSpec,
-    modify,
 )
 from dmr.negotiation import ContentType, conditional_type
 from dmr.openapi import build_schema
 from dmr.openapi.objects import ParameterMetadata
-from dmr.parsers import JsonParser, MultiPartParser
+from dmr.parsers import JsonParser
 from dmr.plugins.pydantic import PydanticSerializer
 from dmr.renderers import JsonRenderer
-from dmr.routing import Router
+from dmr.routing import Router, compose_blueprints
 from dmr.security.jwt import JWTAsyncAuth
 from tests.infra.xml_format import XmlParser, XmlRenderer
 
@@ -34,11 +34,25 @@ class _UserModel(pydantic.BaseModel):
     model_config = pydantic.ConfigDict(title='CustomUserModel')
 
 
-class _UserController(
-    Controller[PydanticSerializer],
+class _PathIdModel(pydantic.BaseModel):
+    id: int
+
+
+class _UserBlueprint(
+    Blueprint[PydanticSerializer],
     Body[dict[str, int]],
 ):
     def post(self) -> _UserModel:
+        raise NotImplementedError
+
+
+class _GetUserListBlueprint(Blueprint[PydanticSerializer]):
+    def get(self) -> list[_UserModel]:
+        raise NotImplementedError
+
+
+class _GetUserController(Controller[PydanticSerializer], Path[_PathIdModel]):
+    def get(self) -> _UserModel:
         raise NotImplementedError
 
 
@@ -49,7 +63,16 @@ def test_user_schema(snapshot: SnapshotAssertion) -> None:
             build_schema(
                 Router(
                     'api/v1/',
-                    [path('user/', _UserController.as_view())],
+                    [
+                        path(
+                            'user/',
+                            compose_blueprints(
+                                _UserBlueprint,
+                                _GetUserListBlueprint,
+                            ).as_view(),
+                        ),
+                        path('user/<int:id>/', _GetUserController.as_view()),
+                    ],
                 ),
             ).convert(),
             indent=2,
@@ -115,82 +138,6 @@ def test_auth_and_cookies_schema(snapshot: SnapshotAssertion) -> None:
     )
 
 
-class _FileModel(pydantic.BaseModel):
-    content_type: Literal['application/json']
-    size: int
-
-
-class _SeveralFiles(pydantic.BaseModel):
-    """Model docs."""
-
-    __dmr_force_list__: ClassVar[frozenset[str]] = frozenset(('attachments',))
-
-    attachments: list[_FileModel]
-    second_file: _FileModel
-
-
-# TODO: test file response
-class _FileController(
-    Controller[PydanticSerializer],
-    FileMetadata[_SeveralFiles],
-):
-    parsers = (MultiPartParser(),)
-
-    @modify(operation_id='file_test_id', deprecated=True)
-    async def get(self) -> list[int]:
-        raise NotImplementedError
-
-
-def test_file_schema(snapshot: SnapshotAssertion) -> None:
-    """Ensure that schema is correct for file controller."""
-    assert (
-        json.dumps(
-            build_schema(
-                Router(
-                    '',
-                    [path('file/', _FileController.as_view())],
-                ),
-            ).convert(),
-            indent=2,
-        )
-        == snapshot
-    )
-
-
-class _DescriptionModel(pydantic.BaseModel):
-    """Description from doc."""
-
-    first: str
-    second: list[int]
-
-
-class _BodyAndFileController(
-    Controller[PydanticSerializer],
-    Body[_DescriptionModel],
-    FileMetadata[_SeveralFiles],
-):
-    parsers = (MultiPartParser(),)
-
-    async def post(self) -> list[int]:
-        raise NotImplementedError
-
-
-def test_body_and_file_schema(snapshot: SnapshotAssertion) -> None:
-    """Ensure that schema is correct for file controller."""
-    assert (
-        json.dumps(
-            build_schema(
-                Router(
-                    '/',
-                    [path('file/', _BodyAndFileController.as_view())],
-                ),
-            ).convert(),
-            indent=2,
-        )
-        == snapshot
-    )
-
-
 class _XmlModel(pydantic.BaseModel):
     xml_value: int
 
@@ -244,6 +191,40 @@ def test_conditional_types(snapshot: SnapshotAssertion) -> None:
                 Router(
                     'api/',
                     [path('types/', _ConditionalTypesController.as_view())],
+                ),
+            ).convert(),
+            indent=2,
+        )
+        == snapshot
+    )
+
+
+class _GetPostController(Controller[PydanticSerializer]):
+    responses = (
+        ResponseSpec(Controller.error_model, status_code=HTTPStatus.NOT_FOUND),
+    )
+
+    def get(self) -> str:
+        raise NotImplementedError
+
+
+def test_raw_path_schema(snapshot: SnapshotAssertion) -> None:
+    """Ensure that schema is correct for raw path items."""
+    assert (
+        json.dumps(
+            build_schema(
+                Router(
+                    'api/v1/',
+                    [
+                        path(
+                            'user/<int:user_id>/post/<uuid:post_id>/',
+                            _GetPostController.as_view(),
+                        ),
+                        re_path(
+                            r'^articles/(?P<year>[0-9]{4})/(?P<slug>[\w-]+)/$',
+                            _GetPostController.as_view(),
+                        ),
+                    ],
                 ),
             ).convert(),
             indent=2,